requirements.txt
apify-client>=1.7.0
requests>=2.31.0
python-dotenv>=1.0.0

# Google integrations
gspread>=5.12.0
google-auth>=2.23.0
google-api-python-client>=2.100.0